+++ R/get_project.R
@@ -14,6 +14,9 @@
 #'   brazil_td_chem      \tab Top-down with chemical mechanisms\tab CETESB\tab  .csv and .rds\cr
 #'   brazil_bu_chem      \tab Bottom-up  chemical mechanisms\tab CETESB+tunnel\tab  .rds\cr
 #'   brazil_bu_chem_streets      \tab Bottom-up  chemical mechanisms for streets and MUNICH\tab CETESB+tunnel\tab  .rds\cr
+#'   brazil_td_chem_im      \tab Top-down with chemical mechanisms+IM\tab CETESB\tab  .csv and .rds\cr
+#'   brazil_bu_chem_im      \tab Bottom-up  chemical mechanisms+IM\tab CETESB+tunnel\tab  .rds\cr
+#'   brazil_bu_chem_streets_im      \tab Bottom-up  chemical mechanisms for streets and MUNICH+IM\tab CETESB+tunnel\tab  .rds\cr
 #'   sebr_cb05co2      \tab Top-down SP, MG and RJ\tab CETESB+tunnel\tab  .rds\cr
 #'   amazon2014      \tab Top-down Amazon\tab CETESB+tunnel\tab  csv and.rds\cr
 #'   curitiba      \tab Bottom-down +GTFS\tab CETESB+tunnel\tab  csv and.rds\cr
@@ -28,8 +31,9 @@
 #' @param url String, with the URL to download VEIN project
 #' @note default case can be any of "brasil", "brazil", "brazil_bu", "brasil_bu", they are
 #' the same
-#' Projects for Ecuador are in development.
-#' In any case, if you find any error, please, send a pull request in github or gitlab.
+#' In any case, if you find any error, please, send a pull request in github.
+#'
+#' In Sao Paulo the IM programs was functioning until 2011. Since then,
 #' @importFrom utils download.file untar
 #' @export
 #' @examples \dontrun{
@@ -92,6 +96,14 @@ get_project <- function(directory,
     utils::untar(tarfile = tf, exdir = directory)
     message("Your directory is in ", directory)
 
+  } else if(case %in% c("brazil_bu_chem_im")){
+    URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/brazil_bu_chem_im.tar.gz"
+    tf <- paste0(tempfile(), ".tar.gz")
+    utils::download.file(url = URL,
+                         destfile =  tf)
+    utils::untar(tarfile = tf, exdir = directory)
+    message("Your directory is in ", directory)
+
   } else if(case %in% c("manizales_bu")){
     URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/manizales_bu.tar.gz"
     tf <- paste0(tempfile(), ".tar.gz")
@@ -108,6 +120,14 @@ get_project <- function(directory,
     utils::untar(tarfile = tf, exdir = directory)
     message("Your directory is in ", directory)
 
+  } else if(case %in% c("brazil_bu_chem_streets_im")){
+    URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/brazil_bu_chem_streets_im.tar.gz"
+    tf <- paste0(tempfile(), ".tar.gz")
+    utils::download.file(url = URL,
+                         destfile =  tf)
+    utils::untar(tarfile = tf, exdir = directory)
+    message("Your directory is in ", directory)
+
   } else if(case %in% c("masp2020")){
     URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/MASP_2020.tar.gz"
     tf <- paste0(tempfile(), ".tar.gz")
@@ -118,6 +138,14 @@ get_project <- function(directory,
 
   } else if(case %in% c("brazil_td_chem")){
     URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/brazil_td_chem.tar.gz"
+    tf <- paste0(tempfile(), ".tar.gz")
+    utils::download.file(url = URL,
+                         destfile =  tf)
+    utils::untar(tarfile = tf, exdir = directory)
+    message("Your directory is in ", directory)
+
+  } else if(case %in% c("brazil_td_chem_im")){
+    URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/brazil_td_chem_im.tar.gz"
     tf <- paste0(tempfile(), ".tar.gz")
     utils::download.file(url = URL,
                          destfile =  tf)